Fix horizontal scrolling
* Add a vbox inside the scrolled_window, and add the two treeviews to that.

This fixes horizontal scrolling. If the budget view is too wide for the
window, then a horizontal scrollbar will appear at the bottom, and scroll
both the main budget view and the totals view at the bottom.

However, it breaks vertical scrolling (vertical scrollbar will scroll the
entire view, instead of just the top view). I'll fix this in the next commit.