Fix vertical scrolling
Vertical scrolling will only apply to the top view (budget values) and not
to the bottom view (budget totals).

Implementation: add an inner scrolled window for the vertical scrollbar.
Set scrolled_window vertical scrollbar policy to NEVER.

Note: This design looks weird when both scrollbars are used because you
have to scroll horizontally to the right in order to see the vertical
scrollbar for the top section.